Rename: data패키지명을 database로 변경 (#8)
